requirements.txt
absl-py==2.1.0
accelerate==1.5.2
aenum==3.1.15
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.4
alembic==1.14.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
anywidget==0.9.13
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beartype==0.20.2
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
byol-pytorch==0.8.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clearml==1.17.2rc0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
crc32c==2.7.1
cycler==0.12.1
databricks-sdk==0.43.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
docker==7.1.0
donfig==0.8.1.post1
efficientnet_pytorch==0.6.3
einops==0.8.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.11.0
fire==0.7.0
Flask==3.1.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
furl==2.1.3
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
imagecodecs==2024.12.30
imageio==2.37.0
imglyb==2.1.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itk==5.4.0
itk-core==5.4.0
itk-filtering==5.4.0
itk-io==5.4.0
itk-numerics==5.4.0
itk-registration==5.4.0
itk-segmentation==5.4.0
itsdangerous==2.2.0
jedi==0.19.2
jgo==1.0.6
Jinja2==3.1.5
joblib==1.4.2
jpype1==1.5.2
json-tricks==3.17.3
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
labeling==0.1.14
lazy_loader==0.4
lightly==1.5.19
lightly-utils==0.0.2
lightning-utilities==0.14.2
lmdb==1.6.2
lpips==0.1.4
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.1
mlflow==2.20.1
mlflow-skinny==2.20.1
monai==1.4.0
mpmath==1.3.0
multidict==6.2.0
munch==4.0.0
narwhals==1.26.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.3
nni==3.0
notebook==7.3.2
notebook_shim==0.2.4
numcodecs==0.15.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ome-types==0.6.1
omegaconf==2.3.0
onnx==1.17.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openslide-python==1.4.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
optuna==4.2.0
orderedmultidict==1.0.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib2==2.3.7.post1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pretrained_microscopy_models @ git+https://github.com/nasa/pretrained-microscopy-models@f50b5689a71cf6f86016fb4a5c5a2bb31e0397f3
pretrainedmodels==0.7.4
prettytable==3.14.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psutil==6.1.1
psygnal==0.12.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyamg==5.2.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic_core==2.27.2
pydicom==3.0.1
Pygments==2.19.1
pyimagej==1.6.0
PyJWT==2.9.0
pynrrd==1.1.3
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
PythonWebHDFS==0.2.3
pytorch-ignite==0.4.11
pytorch-lightning==2.5.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
readlif==0.6.5
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
responses==0.25.6
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
schema==0.7.7
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
scyjava==1.10.2
seaborn==0.13.2
segmentation-models-pytorch==0.2.1
Send2Trash==1.8.3
setuptools==75.8.0
simplejson==3.19.3
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
stringzilla==3.12.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.1.10
timm
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.6.0
torchmetrics==1.7.0
torchview==0.2.6
torchvision==0.21.0
torchviz==0.0.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.2.0
typeguard==4.1.2
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xarray==2025.1.2
xsdata==24.3.1
yarl==1.18.3
zarr==3.0.2
zipp==3.21.0

